time_t is long long on OpenBSD

+++ smtpd/lka_filter.c
@@ -755,13 +755,13 @@ filter_protocol_query(struct filter *filter, uint64_t token, uint64_t reqid, con
 		n = io_printf(lka_proc_get_io(filter->proc),
 		    "filter|%s|%lld.%06ld|smtp-in|%s|%016"PRIx64"|%016"PRIx64"|%s|%s\n",
 		    PROTOCOL_VERSION,
-		    tv.tv_sec, tv.tv_usec,
+		    (long long int)tv.tv_sec, tv.tv_usec,
 		    phase, reqid, token, fs->rdns, param);
 	else
 		n = io_printf(lka_proc_get_io(filter->proc),
 		    "filter|%s|%lld.%06ld|smtp-in|%s|%016"PRIx64"|%016"PRIx64"|%s\n",
 		    PROTOCOL_VERSION,
-		    tv.tv_sec, tv.tv_usec,
+		    (long long int)tv.tv_sec, tv.tv_usec,
 		    phase, reqid, token, param);
 	if (n == -1)
 		fatalx("failed to write to processor");
@@ -779,7 +779,7 @@ filter_data_query(struct filter *filter, uint64_t token, uint64_t reqid, const c
 	    "filter|%s|%lld.%06ld|smtp-in|data-line|"
 	    "%016"PRIx64"|%016"PRIx64"|%s\n",
 	    PROTOCOL_VERSION,
-	    tv.tv_sec, tv.tv_usec,
+	    (long long int)tv.tv_sec, tv.tv_usec,
 	    reqid, token, line);
 	if (n == -1)
 		fatalx("failed to write to processor");
